requirements.txt
streamlit
PyPDF2
openai
pandas
python-docx